Updated memory allocator implementation

+++ src/util/alloc.h
@@ -15,7 +15,7 @@ typedef struct DgBlockInfo {
 typedef int32_t alloch_t;
 
 int32_t DgMakePool(size_t size);
-void DgFreePool(uint16_t index);
-uint16_t DgBestPoolIndex(size_t size);
+void DgFreePool(int32_t index);
+int32_t DgBestPoolIndex(size_t size);
 void *DgAlloc(size_t size);
 void DgFree(void *block);
